Move utility function from metadata into separate utils module
